feat(AITFC-4): offline org-bundle + evidence-pack verification in auths-verifier, exported to WASM
Move the pure offline-verification core down into auths-verifier so every
surface (native CLI, FFI, browser) shares one implementation:
- tlog: RFC 6962 Merkle proof math, checkpoint/proof wire types, and
  TransparencyError, moved from auths-transparency (which now re-exports).
- org_bundle: AirGappedOrgBundle/BundledKel wire types, off-boarding record
  verification, RT-002 bundled-KEL authentication, authority-by-KEL-position
  classification, and verify_org_bundle, moved from auths-sdk (which keeps
  the registry-bound builder and re-exports the rest).
- evidence_pack: pack wire types, per-row authority re-derivation, Merkle
  transparency checks, and verify_evidence_pack_offline, likewise.
- wasm: new verifyOrgBundle / verifyEvidencePackOffline exports — sync,
  panic-free, tagged-union verdicts with 16 MiB input ceilings.

Focused fail-closed error types OrgBundleError (AUTHS-E2201..6) and
EvidencePackError (AUTHS-E2301..3) replace OrgError::Bundle* (E5619-21
retired in registry.lock); error docs regenerated via xtask.

Auths-Id: did:keri:ELv6uW2irGkclnFq8lAsAexmoLwZ-3k-ocwjpFBZsIEG
Auths-Device: did:keri:ELv6uW2irGkclnFq8lAsAexmoLwZ-3k-ocwjpFBZsIEG
