+++ docs/errors/AUTHS-E2201.md
@@ -1,0 +1,9 @@
+# AUTHS-E2201
+
+**Crate:** `auths-verifier`
+
+**Type:** `OrgBundleError::Integrity`
+
+## Message
+
+bundle integrity failure for '{id}': {reason}
