+++ docs/errors/AUTHS-E2202.md
@@ -1,0 +1,9 @@
+# AUTHS-E2202
+
+**Crate:** `auths-verifier`
+
+**Type:** `OrgBundleError::MissingMemberKel`
+
+## Message
+
+bundle is missing the KEL for delegated member '{member}'
