+++ docs/errors/AUTHS-E2203.md
@@ -1,0 +1,9 @@
+# AUTHS-E2203
+
+**Crate:** `auths-verifier`
+
+**Type:** `OrgBundleError::MissingDelegatorSeal`
+
+## Message
+
+member '{member}' has no delegation seal in the org KEL
