+++ docs/errors/AUTHS-E2204.md
@@ -1,0 +1,9 @@
+# AUTHS-E2204
+
+**Crate:** `auths-verifier`
+
+**Type:** `OrgBundleError::Canonicalize`
+
+## Message
+
+canonicalization failed: {0}
