+++ docs/errors/AUTHS-E2205.md
@@ -1,0 +1,9 @@
+# AUTHS-E2205
+
+**Crate:** `auths-verifier`
+
+**Type:** `OrgBundleError::Parse`
+
+## Message
+
+parse failed: {0}
