+++ docs/errors/AUTHS-E2206.md
@@ -1,0 +1,9 @@
+# AUTHS-E2206
+
+**Crate:** `auths-verifier`
+
+**Type:** `OrgBundleError::RecordInvalid`
+
+## Message
+
+offboarding record invalid: {0}
