+++ docs/errors/AUTHS-E2301.md
@@ -1,0 +1,9 @@
+# AUTHS-E2301
+
+**Crate:** `auths-verifier`
+
+**Type:** `EvidencePackError::Canonicalize`
+
+## Message
+
+canonicalization failed: {0}
