+++ docs/errors/AUTHS-E2302.md
@@ -1,0 +1,9 @@
+# AUTHS-E2302
+
+**Crate:** `auths-verifier`
+
+**Type:** `EvidencePackError::Decode`
+
+## Message
+
+decode failed: {0}
