+++ docs/errors/AUTHS-E2303.md
@@ -1,0 +1,9 @@
+# AUTHS-E2303
+
+**Crate:** `auths-verifier`
+
+**Type:** `EvidencePackError::OfflineVerification`
+
+## Message
+
+offline verification failed: {0}
