--- docs/errors/AUTHS-E5619.md
@@ -1,9 +1,0 @@
-# AUTHS-E5619
-
-**Crate:** `auths-sdk`
-
-**Type:** `OrgError::BundleIntegrity`
-
-## Message
-
-bundle integrity failure for '{id}': {reason}
