--- docs/errors/AUTHS-E5620.md
@@ -1,9 +1,0 @@
-# AUTHS-E5620
-
-**Crate:** `auths-sdk`
-
-**Type:** `OrgError::BundleMissingMemberKel`
-
-## Message
-
-bundle is missing the KEL for delegated member '{member}'
