--- docs/errors/AUTHS-E5621.md
@@ -1,9 +1,0 @@
-# AUTHS-E5621
-
-**Crate:** `auths-sdk`
-
-**Type:** `OrgError::BundleMissingDelegatorSeal`
-
-## Message
-
-member '{member}' has no delegation seal in the org KEL
