+++ docs/errors/registry.lock
@@ -55,6 +55,15 @@ AUTHS-E2106 = auths-verifier::CommitVerificationError::HashAlgorithmUnsupported
 AUTHS-E2107 = auths-verifier::CommitVerificationError::SignatureInvalid
 AUTHS-E2108 = auths-verifier::CommitVerificationError::UnknownSigner
 AUTHS-E2109 = auths-verifier::CommitVerificationError::CommitParseFailed
+AUTHS-E2201 = auths-verifier::OrgBundleError::Integrity
+AUTHS-E2202 = auths-verifier::OrgBundleError::MissingMemberKel
+AUTHS-E2203 = auths-verifier::OrgBundleError::MissingDelegatorSeal
+AUTHS-E2204 = auths-verifier::OrgBundleError::Canonicalize
+AUTHS-E2205 = auths-verifier::OrgBundleError::Parse
+AUTHS-E2206 = auths-verifier::OrgBundleError::RecordInvalid
+AUTHS-E2301 = auths-verifier::EvidencePackError::Canonicalize
+AUTHS-E2302 = auths-verifier::EvidencePackError::Decode
+AUTHS-E2303 = auths-verifier::EvidencePackError::OfflineVerification
 AUTHS-E3001 = auths-core::AgentError::KeyNotFound
 AUTHS-E3002 = auths-core::AgentError::IncorrectPassphrase
 AUTHS-E3003 = auths-core::AgentError::MissingPassphrase
@@ -310,9 +319,6 @@ AUTHS-E5615 = auths-sdk::OrgError::IdentityExists
 AUTHS-E5616 = auths-sdk::OrgError::IdentityInit
 AUTHS-E5617 = auths-sdk::OrgError::Attestation
 AUTHS-E5618 = auths-sdk::OrgError::MemberNotDelegable
-AUTHS-E5619 = auths-sdk::OrgError::BundleIntegrity
-AUTHS-E5620 = auths-sdk::OrgError::BundleMissingMemberKel
-AUTHS-E5621 = auths-sdk::OrgError::BundleMissingDelegatorSeal
 AUTHS-E5622 = auths-sdk::OrgError::PolicyCompile
 AUTHS-E5623 = auths-sdk::OrgError::PolicyBlobMissing
 AUTHS-E5624 = auths-sdk::OrgError::PolicyIntegrity
